+++ tests/TypeTest.php
@@ -173,7 +173,7 @@ class TypeTest
         static::assertInstanceOf(NonType::class, $type->nonType);
         static::assertSame('123', $type->nonTypeCallable);
 
-        // NonType will have not access to parent instance because it is a non Type, in fact.
+        // NonType will have no access to parent instance because it is a non Type, in fact.
         static::assertFalse(property_exists($type->nonType, 'parent'));
     }